Update patch set 7
Patch Set 7:

> Why can't we use IA for this?

Because we can't, at least right now? This is LineageParts, not Settings. And even if we could, what would be the benefits (again, LineageParts here)? This just works and we are using our own and known to be good API. Besides we never know when will Google ditch IA for something else.

Patch-set: 7
Label: Verified=0