Update patch set 1
Patch Set 1:

Kanged what I could from the frameworks commit and reworked the rest.

It's shame so much boilerplate is needed just to deal with showing/hiding the led notification but it seems to all work.  If the dialog is open, it correctly turns the led off/on during screen off/on and task switch away/back.

Patch-set: 1
Label: Verified=0