Update patch set 3
Patch Set 3:

Yea, and not only that. When I first put up this patch, this was just a quick way to add the preferences to Settings (via IA). It is not good as is because it doesn't fit properly (if you check that screen, it already contains a "Privacy" preference category). Joey's suggestion was to add the activity directly in Settings xml.

Patch-set: 3
Label: Verified=0