Update patch set 1
Patch Set 1: Code-Review-1

I can see that there is something to fix here but brightness support capability changed when we moved to the sdk so that it's should be available for battery if rgb adjustable is advertised and for notification if either rgb OR brightness is advertised.  In other words, for both cases, brightness controls should be present if rgb is advertised.

Patch-set: 1
Label: Code-Review=-1