Update patch set 5
Patch Set 5:

"Ink" was a pun due to the squid mascot, which made it at least relevant to the post. 

The LLC/Inc thing is going to be more drama than it's worth - back it off to "Lineage" (or "Lingeage" if that typo is intentional).

Patch-set: 5
Reviewer: Abhisek Devkota <848@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0