Update patch set 1
Patch Set 1:

Hmm, is this the right place to fix this?

If a device is not advertising LIGHTS_RGB_BATTERY_LED then we should be ignoring colors altogether in the lineage-sdk lights handling code.

eg wrap this block in if mMultiColorLed  and add an else to set all colors to white otherwise.
https://github.com/LineageOS/android_lineage-sdk/blob/faca0b31b6a2f74c6003a2359691d2f54d37e1eb/sdk/src/java/org/lineageos/internal/notification/LineageBatteryLights.java#L234

Patch-set: 1