Update patch set 1
Patch Set 1:

Let's define, what should adjust these settings?
I assumed that the adjustment will affect the brightness of the LED when charging. However, this does not work. The LED shines in full brightness. Notifications are the same. Set the brightness to 2% (this is a barely noticeable glow). When connecting the charger - full brightness. When the message (Signal or SMS) is received, the LED is in full brightness.

Patch-set: 1
Reviewer: ShevT <9760@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>