Update patch set 1
Patch Set 1: Code-Review-1

Yeah I'm not in favour of this patch as it stands ;-).  I don't think it's fair to require every device to have to implement brightness in their HAL when we can just as easily do it in fw/b (which is of course how it's now /generally/ working today).

I don't think your device is really multicolor RGB.  At least, not in the sense that was originally intended by the lights overlay config (the light colour picker in settings must already be horribly broken for you).

I also like the idea of brightness and colour being separate config bits but the downside is potentially yet another round of overlay updates for every device.  The only way out of that that I can see is to add a capability bits for battery and notification led that mean brightness is not physically adjustable (as opposed to hal adjustable since that's already defined).  But this is now looking pretty messy.

Another idea would be for you to remove multicolor RGB and then work out what we need to add to give you settings features that are of use (and for the light picker to only offer what's possible).  This way, only your devices need updates (and others that fall into this camp; shamu was pretty horrible and is not real adjustable RGB IIRC).   Oh here's a thought - what about new overlay that we use to define the useful range possible for each of your R, G and B ?  Of course it would default to 0-255 so no-one normally need change it.  But you could clamp it according what your device can actually do (and then we figure out how to modify settings and sdk lights to work with limited ranges).

Patch-set: 1
Reviewer: Sam Mortimer <3075@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1