Update patch set 1
Patch Set 1:

> While I appreciate the idea, I'm not a fan of how it's exposed to
 > the user: the Trust interface is not a thing where we throw all the
 > possible stuffs that could weaken the security to the user's face,
 > that's why there are just 4 "core" indicators.
 > 
 > We could add a 5th regarding debugging options, but at the moment
 > it's not presented the right way - let me handle it

Well I'll change the way debugging features are shown because of the reasons in my comment above.
Is that what you thought about when saying it's not presented in the right way?
If not, could you please explain why it is not the right way and how you would like it to be presented?
Moreover, why would you like to do it on your own if we - in my eyes - only need to change a few lines?

However, we should warn the user if the build is signed with test-keys because I think it is crucial for
the security (see https://source.android.com/devices/tech/ota/sign_builds)

Patch-set: 1
Reviewer: Tobias Tefke <18154@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>