Update patch set 2
Patch Set 2:

So this is my evaluation of the proposed addition:

The proposed addition consists in telling whether the user is using a "secure" build by checking the build signature, which from the security pov is really important.

The solution of a "bad status" is not manageable from the os itself as it requires the user to flash a new build.
This situation is supposed to never happen on official builds.

It is impossible for an user to switch from a priv-key signed build to a testkey signed build "by accident" without wiping data or using a special build that allows to change signatures.

It turns out it's really pointless for us to display a persistent information about this "feature" in Trust.

My proposal is the following:

Through the sdk class we display a trust alert (dismissable notification) on boot (as we're already doing with SELinux) whenever the "bad status" is found.
In parts we already have a toggle for disabling trust alerts.

Patch-set: 2
Reviewer: Joey Rizzoli <8461@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>