Update patch set 2
Patch Set 2:

Actually, if we are using notification channels, NotificationRecord#calculateLights() uses the light colors of the channel (which are not set) instead of the notification ones.

And channels do not allow setting the on/off time, so i came up with:


https://review.lineageos.org/c/LineageOS/android_lineage-sdk/+/244516

https://review.lineageos.org/c/LineageOS/android_packages_apps_LineageParts/+/244512


and also for the "always on" feature breaking on one of my (not yet supported) devices, without modifying the lights HAL and/or kernel drivers.


https://review.lineageos.org/c/LineageOS/android_frameworks_base/+/244518

Patch-set: 2
Reviewer: Gerrit User 8328 <8328@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>