Update patch set 3
Patch Set 3:

Sorry I missed this back in Dec (was away travelling at the time), only looking at it now.

I think this is a great feature.  I'm still thinking about where it should live.  Fundamentally, it's a different settings view on the existing wifi/vpn/mobile restrictions and the addition of a "restrict new apps".  To that end, I think it might be better to stick with the preexisting "data restriction" naming theme rather than introducing a new "firewall" concept (which maps to data restrictions).

For example, I'm currently trialling it under data usage as "Manage data restrictions":
https://photos.app.goo.gl/D78uhQMdJh75izHu8

Later on, it might also be nice to expose background data and unrestricted data usage toggles via the same view.  In addition, I have patches locally for a further network restriction - reject all (which I had forgotten about, to be honest).  Perhaps we could (later) make the columns configurable so that you can choose which set of restrictions to view / configure.

Patch-set: 3
CC: Gerrit User 3075 <3075@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>