Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 5: Verified+1 Code-Review-1
> 
> TODO:
> * Uncomment onAppListItemClicked() and have it link to the app data settings.
> * I'm still not keen on "firewall" and would prefer a term like "network restrictions".
> * I would like to wait until https://review.lineageos.org/q/topic:ten-network-isolation is in and then redo the aspect of blocking all access here based on that (instead of blocking wifi,mobile,vpn).

Thank you very much for undertaking the firewall on lineage 17.1! I'm still on 16.0 because of too less spare time the last weeks but I'm glad to see you continued with it.

I renamed the firewall to DataRestriction on 16.0 and also adopted the idea to hide unchangeable system apps and apps which doesn't requested internet permission. 

One problem I encountered with the firewall on 16.0. Maybe it is not a problem on 17.1: 
When you have enabled the restriction of new apps you need to reenable each app you reinstall/update as it get blocked regardless of whether you had it enabled previously or not.

Patch-set: 5
Reviewer: Gerrit User 24774 <24774@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>