ButtonSettings: Rework buttons/keyboard backlight control enablement
Let's end once and for all with the nonsense of using floats
config_keyboardBrightnessSettingDefaultFloat and
config_keyboardKeyboardSettingDefaultFloat to enable button
and keyboard brightness control, respectively.
Instead, use bools config_deviceHasVariableButtonBrightness and
recently added config_deviceHasVariableKeyboardBrightness.

This avoids people with devices with hardware keys having to
overlay config_keyboardBrightnessSettingDefaultFloat or
config_keyboardKeyboardSettingDefaultFloat to a value of 0.0
just to disable the unsupported features.

Change-Id: I0dfd0b5fa51c51398face208f17df70419ef00ca