Update patch set 2
Patch Set 2:

I was going to suggest getSimCarrierId(), but that would break MVNOs it seems...Are we sure it returns an empty/null string when it's not registered? If it returns "0", TextUtils.isEmpty will not help us.

Patch-set: 2
Reviewer: Gerrit User 11801 <11801@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>