Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Why should this be toggable?

it should not, in fact, according to previous comments in the relevant Q commit. I just cherry picked it and tried to make it compile, but did not had the time to remove the toggable part. Feel free to remove it, if you have some spare time :)

Patch-set: 1