Update patch set 1
Patch Set 1:

Hi Sam!

For me it is ok to put it in the network settings as it changes network settings ;)

You're right. VPN isn't used that much and enabling/disabling all network access of an app is more useful. I think we also should enable a horizontal scroll bar for narrower devices when too much checkboxes are enabled.

I will write here when I have enough time to have a look on this. Maybe next week.

Thanks,
Stefano

Patch-set: 1