LightSettingsDialog: Do not call super on onStart/onStop
There is no need for this in such custom AlertDialog and it avoids triggering an IllegalStateException at least during stop, since the view lifecycle is in INITIALIZED state, never changes
to CREATED and therefore can't be moved to DESTROYED from that
initial state.

Fixes app crashing when applying / dismissing changes in notification
light settings.

Change-Id: I21054cde355582a0a55309850a66f63061b917d3